backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
opencv-python>=4.9.0
pillow>=10.0.0
grad-cam>=1.5.4
pydantic>=2.8.0
requests>=2.32.0
python-dotenv>=1.0.1
torch
torchvision
